WebAPI

+++ Entity/Concrete/Car.cs
@@ -13,7 +13,7 @@ namespace Entity.Concrete
         public int BrandId { get; set; }
         public int ColorId { get; set; }
         public short ModelYear { get; set; }
-        public decimal DailyPrice { get; set; }
+        public double DailyPrice { get; set; }
         public string Description { get; set; }
 
     }
