+++ Entity/DTOs/CarDetailDto.cs
@@ -10,6 +10,6 @@ namespace Entity.DTOs
         public int CarId { get; set; }
         public string BrandName { get; set; }
         public string ColorName { get; set; }
-        public decimal DailyPrice { get; set; }
+        public double DailyPrice { get; set; }
     }
 }
